+++ example/metro.config.js
@@ -2,7 +2,6 @@ const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
 
 const path = require('path');
 const escape = require('escape-string-regexp');
-const exclusionList = require('metro-config/src/defaults/exclusionList');
 const pak = require('../package.json');
 
 const root = path.resolve(__dirname, '..');
@@ -10,6 +9,32 @@ const root = path.resolve(__dirname, '..');
 const modules = Object.keys({
   ...pak.peerDependencies,
 });
+
+// exclusionList function copied from metro-config
+const list = [/\/__tests__\/.*/];
+function escapeRegExp(pattern) {
+  if (pattern instanceof RegExp) {
+    return pattern.source.replace(/\/|\\\//g, '\\' + path.sep);
+  } else if (typeof pattern === 'string') {
+    const escaped = pattern.replace(
+      /[\-\[\]\{\}\(\)\*\+\?\.\\\^\$\|]/g,
+      '\\$&'
+    );
+    return escaped.replaceAll('/', '\\' + path.sep);
+  } else {
+    throw new Error(
+      `Expected exclusionList to be called with RegExp or string, got: ${typeof pattern}`
+    );
+  }
+}
+function exclusionList(additionalExclusions) {
+  return new RegExp(
+    '(' +
+      (additionalExclusions || []).concat(list).map(escapeRegExp).join('|') +
+      ')$'
+  );
+}
+// exclusionList end
 
 /**
  * Metro configuration